refactored User to have has settings trait
