+++ config/mail.php
@@ -103,7 +103,7 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Markdown Mail Settings
+    | Markdown Mail hasSettings
     |--------------------------------------------------------------------------
     |
     | If you are using Markdown based email rendering, you may configure your
