+++ routes/api.php
@@ -21,7 +21,7 @@ Route::get('/user', 'AuthController@index')->middleware('auth:api')->name('curre
 // Admin
 Route::patch('/users/{user}/admin' , 'AdminController@update')->middleware('auth:api' , 'admin')->name('updateUser');
 
-// Settings
+// hasSettings
 Route::post('/users/{user}/settings' , 'SettingsController@store')->middleware('auth:api')->name('settings');
 
 Route::middleware(['auth:api'])->group(function () {
